Fix `is_instantiated()` to check for all artifacts
Previously, we had issues with `is_instantiated()` not checking that all
artifacts were downloaded.  This is because by iterating over the
manifest in `ctx.env`, we only check for `Artifacts.toml` files within
dependencies, not within the top-level project.  This commit alters the
instantiation checks to also check the top-level project for
`Artifacts.toml` files that may need to be downloaded as well.